gather/requirements.txt
beautifulsoup4==4.10.0
requests==2.26.0
boto3==1.20.13
python-dotenv == 0.19.0
